int cast

+++ Libft/ft_strrstr.c
@@ -6,7 +6,7 @@
 /*   By: hel-makh <hel-makh@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/05 14:47:08 by hel-makh          #+#    #+#             */
-/*   Updated: 2022/06/05 17:32:57 by hel-makh         ###   ########.fr       */
+/*   Updated: 2022/06/20 18:39:15 by hel-makh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,12 @@ char	*ft_strrstr(const char *haystack, const char *needle)
 	while (i >= 0)
 	{
 		j = ft_strlen(needle) - 1;
-		while (i - (ft_strlen(needle) - 1 - j) >= 0
-			&& haystack[i - (ft_strlen(needle) - 1 - j)] == needle[j])
+		while (i - ((int)ft_strlen(needle) - 1 - j) >= 0
+			&& haystack[i - ((int)ft_strlen(needle) - 1 - j)] == needle[j])
 		{
 			if (j == 0)
-				return ((char *)&haystack[i - (ft_strlen(needle) - 1 - j)]);
+				return ((char *)&haystack
+					[i - ((int)ft_strlen(needle) - 1 - j)]);
 			j --;
 		}
 		i --;
